+++ CHANGELOG.md
@@ -1,6 +1,8 @@
+# 1.2.0
+* Adding a KeyToValue transformation to allow for key to be stored in a separate column in ClickHouse
+
 # 1.1.4
 * Bugfix to address field value to column name mapping for Tuples
-* Adding a KeyToValue transformation to allow for key to be stored in a separate column in ClickHouse
 
 # 1.1.3
 * Update to java-client 0.6.3